createアクション

+++ app/controllers/public/orders_controller.rb
@@ -16,10 +16,19 @@ class Public::OrdersController < ApplicationController
   def create
     @order = Order.new(order_params)
     @order.customer_id = current_customer.id
-    @order.save
-    
-    current_customer.cart_items.destroy_all
-    redirect_to public_orders_complete_path
+    if @order.save
+      @cart_items = current_customer.cart_items
+      @cart_items.each do |cart_item|
+        @order_detail = OrderDetail.new
+        @order_detail.order_id = @order.id
+        @order_detail.item_id = cart_item.item_id
+        @order_detail.price = cart_item.item.prise
+        @order_detail.amount = cart_item.amount
+        @order_detail.save
+      end
+      current_customer.cart_items.destroy_all
+      redirect_to public_orders_complete_path
+    end
   end
 
   def confirm
@@ -46,7 +55,7 @@ class Public::OrdersController < ApplicationController
    private
 
   def order_params
-    params.require(:order).permit(:name, :postal_code, :address, :total_payment, :payment_method)
+    params.require(:order).permit(:name, :postal_code, :address, :total_payment, :payment_method, :shipping_cost)
   end
 
 end
